Move get_vendor_name() to util.py
